refactor: repository 인터페이스 삭제
