--- src/main/java/roomescape/time/service/ReservationTimeRepository.java
@@ -1,16 +1,0 @@
-package roomescape.time.service;
-
-import java.util.List;
-import java.util.Optional;
-import roomescape.time.domain.ReservationTime;
-
-public interface ReservationTimeRepository {
-
-    ReservationTime save(ReservationTime reservationTime);
-
-    List<ReservationTime> findAll();
-
-    void deleteById(Long id);
-
-    Optional<ReservationTime> findById(Long id);
-}
